restructured CRA
