+++ src/_components/About/About.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const About = () => {
+  return (
+    <div>
+      <h1>Footer</h1>
+    </div>
+  );
+};
+
+export { About };
