+++ src/_components/Footer/Footer.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Footer = () => {
+  return (
+    <div>
+      <h1>Footer</h1>
+    </div>
+  );
+};
+
+export { Footer };
