+++ src/_components/Hero/Hero.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Hero = () => {
+  return (
+    <div>
+      <h1>Hero</h1>
+    </div>
+  );
+};
+
+export { Hero };
