+++ src/_components/Navbar/Navbar.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Navbar = () => {
+  return (
+    <div>
+      <h1>Navbar</h1>
+    </div>
+  );
+};
+
+export { Navbar };
