+++ src/_components/Projects/Projects.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Projects = () => {
+  return (
+    <div>
+      <h1>Projects</h1>
+    </div>
+  );
+};
+
+export { Projects };
