+++ src/_components/Projects/StaticProjectGallery.js
@@ -1,0 +1,14 @@
+// This component is a hard-coded project section so that if
+// the server is down, the project section of the portfolio
+// will not break
+import React from "react";
+
+const StaticProjectGallery = () => {
+  return (
+    <ProjectGallery>
+      <h1>Static Project Gallery</h1>
+    </ProjectGallery>
+  );
+};
+
+export default StaticProjectGallery;
